+++ src/vue/navigation/Footer.vue
@@ -5,6 +5,7 @@
       <router-link
         class="footer__link"
         :to="vueRoutes.terms"
+        target="_blank"
       >
         {{ 'footer.terms' | globalize }}
       </router-link>
@@ -12,6 +13,7 @@
       <router-link
         class="footer__link"
         :to="vueRoutes.downloads"
+        target="_blank"
       >
         {{ 'footer.download-apps' | globalize }}
       </router-link>
@@ -19,6 +21,7 @@
       <a
         class="footer__link"
         :href="config.SUPPORT_URL"
+        target="_blank"
       >
         {{ 'footer.support-apps' | globalize }}
       </a>